Extract the test
